+++ example/lib/main.dart
@@ -8,9 +8,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Expand Widget',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
+      theme: ThemeData(primarySwatch: Colors.blue),
       home: HomePage(),
     );
   }
@@ -20,9 +18,7 @@ class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text('Expand Widget'),
-      ),
+      appBar: AppBar(title: Text('Expand Widget')),
       body: ListView(
         padding: EdgeInsets.all(8),
         children: <Widget>[
@@ -55,7 +51,7 @@ Curabitur fringilla enim at lorem pulvinar, id ornare tellus aliquam. Cras eget 
               child: Column(
                 children: <Widget>[
                   Text(
-                    'Expand Widgets',
+                    'Expand Child',
                     style: Theme.of(context).textTheme.title,
                   ),
                   SizedBox(height: 8),
@@ -64,6 +60,51 @@ Curabitur fringilla enim at lorem pulvinar, id ornare tellus aliquam. Cras eget 
                     onPressed: () => print('Pressed button0'),
                   ),
                   ExpandChild(
+                    child: Column(
+                      children: <Widget>[
+                        OutlineButton(
+                          child: Text('Button1'),
+                          onPressed: () => print('Pressed button1'),
+                        ),
+                        OutlineButton(
+                          child: Text('Button2'),
+                          onPressed: () => print('Pressed button2'),
+                        ),
+                        OutlineButton(
+                          child: Text('Button3'),
+                          onPressed: () => print('Pressed button3'),
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+          Card(
+            child: Padding(
+              padding: EdgeInsets.all(8),
+              child: Column(
+                children: <Widget>[
+                  Text(
+                    'Show Child',
+                    style: Theme.of(context).textTheme.title,
+                  ),
+                  SizedBox(height: 8),
+                  OutlineButton(
+                    child: Text('Button0'),
+                    onPressed: () => print('Pressed button0'),
+                  ),
+                  ShowChild(
+                    indicator: Padding(
+                      padding: EdgeInsets.all(8),
+                      child: Text(
+                        "SHOW MORE",
+                        style: TextStyle(
+                          color: Theme.of(context).textTheme.caption.color,
+                        ),
+                      ),
+                    ),
                     child: Column(
                       children: <Widget>[
                         OutlineButton(