+++ lib/expand_widget.dart
@@ -2,3 +2,4 @@ library expand_widget;
 
 export 'src/expand_child.dart';
 export 'src/expand_text.dart';
+export 'src/show_child.dart';