+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+## [0.2.0] - June 24th, 2019
+
+* Added `ShowChild` widget.
+
 ## [0.1.2] - June 8th, 2019
 
-* Default arrow size changed to 27
-* Updated example project
+* Default arrow size changed to 27.
+* Updated example project.
 
 ## [0.1.1] - June 7th, 2019
 
-* Increased `maxLenght` from 5 to 8
+* Increased `maxLenght` from 5 to 8.
 
 ## [0.1.0] - June 7th, 2019